test: 사람의 이름이 5자를 초과할 경우 예외가 발생한다.
