--- src/main/java/nextstep/ladder/User.java
@@ -1,13 +1,0 @@
-package nextstep.ladder;
-
-public class User {
-    private final String name;
-
-    public User(String name) {
-        this.name = name;
-    }
-
-    public String getName() {
-        return name;
-    }
-}
